requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.35
alembic>=1.13.3
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
jinja2>=3.1.4
python-dotenv>=1.0.1
httpx>=0.27.2
itsdangerous>=2.2.0
aiofiles>=24.1.0
